refactor: rename components and update styles for SpecialSuggest and PopularDestinations; adjust body overflow handling in Header
