+++ src/app/page.jsx
@@ -1,6 +1,6 @@
 import Counter from "@/layout/Counter";
-import SpecialSuggest from "@/layout/SpecialSuggest";
-import PopularDestinations from "@/layout/PopularDestinations";
+import SpecialSuggest from "@/layout/PopularDestinations";
+import PopularDestinations from "@/layout/SpecialSuggest";
 import Cta from "@/layout/Cta";
 import Features from "@/layout/Features";
